fix: have the source files included in the 'package' command's output
Dependencies added to yarn using git protocol will run `yarn pack` on
the collected files first. Currently, `files` field in the package.json
instructs `yarn pack` to only include 'dist' folder, which is not
available in the repo since they are output of the build process and
hence set to be ignored using `.gitignore`.

Since we're building this from the source when running `install` in
our project, having source files is essential. Only including 'dist'
will cause the build to fail during installation.

Signed-off-by: zyrif <rubaiyat9@gmail.com>